+++ mybatis-plus-dts/src/main/java/com/baomidou/mybatisplus/dts/EnableDtsRabbit.java
@@ -1,7 +1,6 @@
 package com.baomidou.mybatisplus.dts;
 
-import com.baomidou.mybatisplus.dts.rabbit.config.RmtAutoConfiguration;
-import com.baomidou.mybatisplus.dts.rabbit.mq.RabbitConfiguration;
+import com.baomidou.mybatisplus.dts.rabbit.RabbitConfiguration;
 import org.springframework.context.annotation.Import;
 
 import java.lang.annotation.Documented;
@@ -20,7 +19,7 @@ import java.lang.annotation.Target;
 @Retention(value = java.lang.annotation.RetentionPolicy.RUNTIME)
 @Target(value = { java.lang.annotation.ElementType.TYPE })
 @Documented
-@Import({RabbitConfiguration.class, RmtAutoConfiguration.class})
-public @interface EnableRmtRabbit {
+@Import({RabbitConfiguration.class})
+public @interface EnableDtsRabbit {
 
 }
