+++ mybatis-plus-dts/src/main/java/com/baomidou/mybatisplus/dts/parser/JacksonRmtParser.java
@@ -1,4 +1,4 @@
-package com.baomidou.mybatisplus.dts.rabbit.parser;
+package com.baomidou.mybatisplus.dts.parser;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
